social_networking/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==1.7.1
sqlparse==0.5.0
typing_extensions==4.12.0
